pyproject.toml
[project]
name = "nvr"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi",
    "pyyaml",
    "uvicorn[standard]",
]

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py314"

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "UP", "B", "A", "C4",
    "SIM", "PL", "RUF", "PIE", "RET", "PERF", "FURB",
    "LOG", "G", "T20", "PTH", "TID",
]
ignore = ["PLR09", "PLR2004", "PLW2901", "SIM108"]

[tool.pyright]
pythonVersion = "3.14"
venvPath = "."
venv = ".venv"
